pyproject.toml
[tool.poetry]
name = "rotate-files"
version = "0.1.0"
description = ""
authors = ["Ties de Kock <ties@tiesdekock.nl>"]
readme = "README.md"
packages = [{include = "rotate_files"}]

[tool.poetry.dependencies]
python = "^3.11"
PyYAML = "^6.0.1"
click = "^8.1.6"
aiohttp = "^3.8.5"
pytest-asyncio = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
tox = "^4.4.12"
isort = {extras = ["pyproject"], version = "^5.12.0"}
pytest-cov = "^4.0.0"
pytest-xdist = "^3.2.1"
ruff = "^0.0.261"
black = "^23.3.0"
twine = "^4.0.2"
check-manifest = "^0.49"
flake8 = "^6.0.0"
pytest-aiohttp = "^1.0.4"

[tool.black]
target-version = ["py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88
profile = "black"

[tool.ruff]
ignore = ["E501"]

[tool.pytest]
junit_family = "xunit2"

[settings]
# known_third_party = []